+++ src/protocols/on_disconnect.rs
@@ -19,11 +19,11 @@ use crate::{
 /// from the associated connection (i.e. [`Reading`] is enabled) or if it attempts to send a message
 /// to it (i.e. one of the [`Writing`] methods is called).
 #[async_trait::async_trait]
-pub trait Disconnect: Pea2Pea
+pub trait OnDisconnect: Pea2Pea
 where
     Self: Clone + Send + Sync + 'static,
 {
-    /// Attaches the behavior specified in [`Disconnect::handle_disconnect`] to every occurrence of the
+    /// Attaches the behavior specified in [`OnDisconnect::on_disconnect`] to every occurrence of the
     /// node disconnecting from a peer.
     async fn enable_disconnect(&self) {
         let (from_node_sender, mut from_node_receiver) =
@@ -35,9 +35,9 @@ where
         // spawn a background task dedicated to handling disconnect events
         let self_clone = self.clone();
         let disconnect_task = tokio::spawn(async move {
-            trace!(parent: self_clone.node().span(), "spawned the Disconnect handler task");
+            trace!(parent: self_clone.node().span(), "spawned the OnDisconnect handler task");
             if tx.send(()).is_err() {
-                error!(parent: self_clone.node().span(), "Disconnect handler creation interrupted! shutting down the node");
+                error!(parent: self_clone.node().span(), "OnDisconnect handler creation interrupted! shutting down the node");
                 self_clone.node().shut_down().await;
                 return;
             }
@@ -46,7 +46,7 @@ where
                 let self_clone2 = self_clone.clone();
                 tokio::spawn(async move {
                     // perform the specified extra actions
-                    self_clone2.handle_disconnect(addr).await;
+                    self_clone2.on_disconnect(addr).await;
                     // notify the node that the extra actions have concluded
                     // and that the related connection can be dropped
                     let _ = notifier.send(()); // can't really fail
@@ -57,18 +57,18 @@ where
         self.node()
             .tasks
             .lock()
-            .insert(NodeTask::Disconnect, disconnect_task);
+            .insert(NodeTask::OnDisconnect, disconnect_task);
 
-        // register the Disconnect handler with the Node
+        // register the OnDisconnect handler with the Node
         let hdl = ProtocolHandler(from_node_sender);
         assert!(
-            self.node().protocols.disconnect.set(hdl).is_ok(),
-            "the Disconnect protocol was enabled more than once!"
+            self.node().protocols.on_disconnect.set(hdl).is_ok(),
+            "the OnDisconnect protocol was enabled more than once!"
         );
     }
 
     /// Any extra actions to be executed during a disconnect; in order to still be able to
     /// communicate with the peer in the usual manner (i.e. via [`Writing`]), only its [`SocketAddr`]
     /// (as opposed to the related [`Connection`] object) is provided as an argument.
-    async fn handle_disconnect(&self, addr: SocketAddr);
+    async fn on_disconnect(&self, addr: SocketAddr);
 }